// generated/src/app/api-models/auth-waive-deter-values.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuthWaiveDeterValues {
  seqAwRule: number;
  searchSequence: number;
  determinantSequence: number;
  detFromValue: string;
  detThruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}